#175 Refactor: 공유 서비스 통합
