+++ linknamu/src/main/java/com/kakao/linknamu/share/controller/ShareController.java
@@ -5,12 +5,8 @@ import com.kakao.linknamu.core.util.ApiUtils;
 import com.kakao.linknamu.share.dto.category.CreateCategoryFromEncodedIdRequestDto;
 import com.kakao.linknamu.share.dto.category.GetCategoryFromLinkResponseDto;
 import com.kakao.linknamu.share.dto.workspace.GetWorkSpaceFromLinkResponseDto;
-import com.kakao.linknamu.share.service.category.CreateCategoryFromEncodedIdService;
-import com.kakao.linknamu.share.service.category.CreateLinkFromCategoryService;
-import com.kakao.linknamu.share.service.category.GetCategoryFromEncodedIdService;
-import com.kakao.linknamu.share.service.workspace.CreateLinkFromWorkspaceService;
-import com.kakao.linknamu.share.service.workspace.CreateWorkspaceFromEncodedIdService;
-import com.kakao.linknamu.share.service.workspace.GetWorkspaceFromEncodedIdService;
+import com.kakao.linknamu.share.service.category.ShareCategoryService;
+import com.kakao.linknamu.share.service.workspace.ShareWorkspaceService;
 import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
@@ -27,25 +23,21 @@ import org.springframework.web.bind.annotation.*;
 @Validated
 public class ShareController {
 	private static final int PAGE_SIZE = 10;
-	private final CreateLinkFromWorkspaceService createLinkFromWorkspaceService;
-	private final GetWorkspaceFromEncodedIdService getWorkspaceFromEncodedIdService;
-	private final CreateWorkspaceFromEncodedIdService createWorkspaceFromEncodedIdService;
-	private final CreateLinkFromCategoryService createLinkFromCategoryService;
-	private final GetCategoryFromEncodedIdService getCategoryFromEncodedIdService;
-	private final CreateCategoryFromEncodedIdService createCategoryFromEncodedIdService;
+	private final ShareCategoryService shareCategoryService;
+	private final ShareWorkspaceService shareWorkspaceService;
 
 	// 워크스페이스
 	@GetMapping("/workspace/{workSpaceId}")
 	public ResponseEntity<?> createLinkFromWorkSpaceId(
-		@PathVariable("workSpaceId") @Positive(message = "id는 양수여야함.") Long workSpaceId) {
-		String link = createLinkFromWorkspaceService.createLink(workSpaceId);
+		@PathVariable("workSpaceId") @Positive(message = "id는 양수여야한다.") Long workSpaceId) {
+		String link = shareWorkspaceService.createLink(workSpaceId);
 		return ResponseEntity.ok(ApiUtils.success(link));
 	}
 
 	@GetMapping("/workspace/link/{encodedWorkSpaceId}")
 	public ResponseEntity<?> getWorkspaceFromEncodedId(
 		@PathVariable String encodedWorkSpaceId) {
-		GetWorkSpaceFromLinkResponseDto responseDto = getWorkspaceFromEncodedIdService.getWorkspace(encodedWorkSpaceId);
+		GetWorkSpaceFromLinkResponseDto responseDto = shareWorkspaceService.getWorkspace(encodedWorkSpaceId);
 		return ResponseEntity.ok(ApiUtils.success(responseDto));
 	}
 
@@ -53,15 +45,15 @@ public class ShareController {
 	public ResponseEntity<?> createWorkspaceFromEncodedId(
 		@PathVariable String encodedWorkSpaceId,
 		@AuthenticationPrincipal CustomUserDetails userDetails) {
-		createWorkspaceFromEncodedIdService.createWorkSpace(encodedWorkSpaceId, userDetails.getUser());
+		shareWorkspaceService.createWorkSpace(encodedWorkSpaceId, userDetails.getUser());
 		return ResponseEntity.ok(ApiUtils.success(null));
 	}
 
 	// 카테고리
 	@GetMapping("/category/{categoryId}")
 	public ResponseEntity<?> createLinkFromCategoryId(
-		@PathVariable("categoryId") @Positive(message = "id는 양수여야함.") Long categoryId) {
-		String link = createLinkFromCategoryService.createLink(categoryId);
+		@PathVariable("categoryId") @Positive(message = "id는 양수여야한다.") Long categoryId) {
+		String link = shareCategoryService.createLink(categoryId);
 		return ResponseEntity.ok(ApiUtils.success(link));
 	}
 
@@ -69,7 +61,7 @@ public class ShareController {
 	public ResponseEntity<?> getCategoryFromEncodedId(
 		@PathVariable String encodedCategoryId, @RequestParam(defaultValue = "0") int page) {
 		Pageable pageable = PageRequest.of(page, PAGE_SIZE, Sort.by("createdAt").descending());
-		GetCategoryFromLinkResponseDto responseDto = getCategoryFromEncodedIdService.getCategory(encodedCategoryId,
+		GetCategoryFromLinkResponseDto responseDto = shareCategoryService.getCategory(encodedCategoryId,
 			pageable);
 		return ResponseEntity.ok(ApiUtils.success(responseDto));
 	}
@@ -79,7 +71,7 @@ public class ShareController {
 		@PathVariable String encodedCategoryId,
 		@RequestBody CreateCategoryFromEncodedIdRequestDto requestDto,
 		@AuthenticationPrincipal CustomUserDetails userDetails) {
-		createCategoryFromEncodedIdService.createCategory(encodedCategoryId, requestDto, userDetails.getUser());
+		shareCategoryService.createCategory(encodedCategoryId, requestDto, userDetails.getUser());
 		return ResponseEntity.ok(ApiUtils.success(null));
 	}
